Add working multi-objective optimization loop, optimizes around U1 and S

+++ coam/templates/clamp.py
@@ -1,3 +1,5 @@
+import json
+
 import visualization
 from abaqus import *
 from abaqusConstants import *
@@ -16,7 +18,7 @@ mdb.models.changeKey(fromName="Model-1", toName="ClampSimulation")
 model = mdb.models["ClampSimulation"]
 
 jaw_actuated_acis = session.openStep(
-    "{{ cwd }}/iterations/{{ iteration }}/jaw_actuated.step",
+    "jaw_actuated.step",
     scaleFromFile=OFF,
 )
 jaw_actuated = model.PartFromGeometryFile(
@@ -29,7 +31,7 @@ jaw_actuated = model.PartFromGeometryFile(
 jaw_actuated = model.parts["jaw_actuated"]
 
 jaw_fixed_acis = session.openStep(
-    "{{ cwd }}/iterations/{{ iteration }}/jaw_fixed.step",
+    "jaw_fixed.step",
     scaleFromFile=OFF,
 )
 jaw_fixed = model.PartFromGeometryFile(
@@ -42,7 +44,7 @@ jaw_fixed = model.PartFromGeometryFile(
 jaw_fixed = model.parts["jaw_fixed"]
 
 part_geometry_acis = session.openStep(
-    "{{ cwd }}/iterations/{{ iteration }}/part_geometry_simplified.step",
+    "part_geometry_simplified.step",
     scaleFromFile=OFF,
 )
 part_geometry = model.PartFromGeometryFile(
@@ -54,35 +56,39 @@ part_geometry = model.PartFromGeometryFile(
 )
 part_geometry = model.parts["part_geometry"]
 
-reference_point_load = jaw_actuated.ReferencePoint(point=(-10.0,
-    0.0, 0.0))
-reference_point_static = jaw_fixed.ReferencePoint(point=(10.0,
-    0.0, 0.0))
+reference_point_load = jaw_actuated.ReferencePoint(point=(-10.0, 0.0, 0.0))
+reference_point_static = jaw_fixed.ReferencePoint(point=(10.0, 0.0, 0.0))
 
-jaw_actuated.Surface(name='Surface',
-    side1Faces=
-    jaw_actuated.faces)
-jaw_fixed.Surface(name='Surface',
-    side1Faces=
-    jaw_fixed.faces)
-part_geometry.Surface(name='Surface',
-    side1Faces=
-    part_geometry.faces)
+jaw_actuated.Surface(name="Surface", side1Faces=jaw_actuated.faces)
+jaw_fixed.Surface(name="Surface", side1Faces=jaw_fixed.faces)
+part_geometry.Surface(name="Surface", side1Faces=part_geometry.faces)
 
 if os.path.exists("retry.flag"):
     try:
         jaw_fixed.createVirtualTopology(
-                applyBlendControls=False, cornerAngleTolerance=30.0,
-                ignoreRedundantEntities=False, mergeShortEdges=False, mergeSliverFaces=
-                False, mergeSmallAngleFaces=False, mergeSmallFaces=True,
-                mergeThinStairFaces=False, smallFaceAreaThreshold=0.1)
+            applyBlendControls=False,
+            cornerAngleTolerance=30.0,
+            ignoreRedundantEntities=False,
+            mergeShortEdges=False,
+            mergeSliverFaces=False,
+            mergeSmallAngleFaces=False,
+            mergeSmallFaces=True,
+            mergeThinStairFaces=False,
+            smallFaceAreaThreshold=0.1,
+        )
         jaw_actuated.createVirtualTopology(
-                applyBlendControls=False, cornerAngleTolerance=30.0,
-                ignoreRedundantEntities=False, mergeShortEdges=False, mergeSliverFaces=
-                False, mergeSmallAngleFaces=False, mergeSmallFaces=True,
-                mergeThinStairFaces=False, smallFaceAreaThreshold=0.1)
-    except Exception, e:
-          pass
+            applyBlendControls=False,
+            cornerAngleTolerance=30.0,
+            ignoreRedundantEntities=False,
+            mergeShortEdges=False,
+            mergeSliverFaces=False,
+            mergeSmallAngleFaces=False,
+            mergeSmallFaces=True,
+            mergeThinStairFaces=False,
+            smallFaceAreaThreshold=0.1,
+        )
+    except Exception as e:
+        pass
 
 additive_steel_material = model.Material(name="Stainless_Steel_316L")
 additive_steel_material.setValues(description="")
@@ -155,8 +161,16 @@ part_geometry.SectionAssignment(
     thicknessAssignment=FROM_SECTION,
 )
 
-jaw_actuated.RemoveCells(cellList=jaw_actuated.cells)
-jaw_fixed.RemoveCells(cellList=jaw_fixed.cells)
+try:
+    jaw_actuated.RemoveCells(cellList=jaw_actuated.cells)
+except Exception as e:
+    pass
+
+try:
+    jaw_fixed.RemoveCells(cellList=jaw_fixed.cells)
+except Exception as e:
+    pass
+
 
 part_geometry.setMeshControls(
     elemShape=TET,
@@ -164,13 +178,10 @@ part_geometry.setMeshControls(
     technique=FREE,
 )
 part_geometry.setElementType(
-    elemTypes=(
-        ElemType(elemCode=C3D20R, elemLibrary=STANDARD),
-        ElemType(elemCode=C3D15, elemLibrary=STANDARD),
-        ElemType(elemCode=C3D10, elemLibrary=STANDARD),
-    ),
+    elemTypes=(ElemType(elemCode=C3D10, elemLibrary=STANDARD),),
     regions=part_set,
 )
+
 part_geometry.seedPart(deviationFactor=0.1, minSizeFactor=0.1, size=4)
 part_geometry.generateMesh()
 
@@ -180,16 +191,22 @@ jaw_actuated.generateMesh()
 jaw_fixed.seedPart(deviationFactor=0.1, minSizeFactor=0.1, size=500)
 jaw_fixed.generateMesh()
 
-part_instance = model.rootAssembly.Instance(dependent=ON, name="part_geometry-1", part=part_geometry)
-jaw_actuated_instance = model.rootAssembly.Instance(dependent=ON, name="jaw_actuated-1", part=jaw_actuated)
-jaw_fixed_instance = model.rootAssembly.Instance(dependent=ON, name="jaw_fixed-1", part=jaw_fixed)
-jaw_actuated_instance_set = model.rootAssembly.Set(name="jaw_actuated_Set",
-    referencePoints=(
-    jaw_actuated_instance.referencePoints[2], ))
+part_instance = model.rootAssembly.Instance(
+    dependent=ON, name="part_geometry-1", part=part_geometry
+)
+jaw_actuated_instance = model.rootAssembly.Instance(
+    dependent=ON, name="jaw_actuated-1", part=jaw_actuated
+)
+jaw_fixed_instance = model.rootAssembly.Instance(
+    dependent=ON, name="jaw_fixed-1", part=jaw_fixed
+)
+jaw_actuated_instance_set = model.rootAssembly.Set(
+    name="jaw_actuated_Set", referencePoints=(jaw_actuated_instance.referencePoints[2],)
+)
 
-jaw_fixed_instance_set = model.rootAssembly.Set(name="jaw_fixed_Set",
-    referencePoints=(
-    jaw_fixed_instance.referencePoints[2], ))
+jaw_fixed_instance_set = model.rootAssembly.Set(
+    name="jaw_fixed_Set", referencePoints=(jaw_fixed_instance.referencePoints[2],)
+)
 
 model.rootAssembly.regenerate()
 model.StaticStep(
@@ -257,32 +274,38 @@ model.interactionProperties["IntProp-Fric"].NormalBehavior(
     allowSeparation=ON, constraintEnforcementMethod=DEFAULT, pressureOverclosure=HARD
 )
 
-global_contact = model.ContactStd(createStepName='Initial', name=
-'GlobalContact')
-global_contact.setValues(globalSmoothing=
-                         False)
+global_contact = model.ContactStd(createStepName="Initial", name="GlobalContact")
+global_contact.setValues(globalSmoothing=False)
 global_contact.contactPropertyAssignments.appendInStep(
-    assignments=((GLOBAL, SELF, 'IntProp-Fric'),), stepName='Initial')
+    assignments=((GLOBAL, SELF, "IntProp-Fric"),), stepName="Initial"
+)
 global_contact.includedPairs.setValuesInStep(
-    addPairs=((
-    jaw_actuated_instance.surfaces['Surface'],
-    part_instance.surfaces['Surface']),
-    (
-    jaw_fixed_instance.surfaces['Surface'],
-    part_instance.surfaces['Surface']))
-    , stepName='Initial', useAllstar=OFF)
-model.StdInitialization(name='ContactInitialization')
+    addPairs=(
+        (jaw_actuated_instance.surfaces["Surface"], part_instance.surfaces["Surface"]),
+        (jaw_fixed_instance.surfaces["Surface"], part_instance.surfaces["Surface"]),
+    ),
+    stepName="Initial",
+    useAllstar=OFF,
+)
+model.StdInitialization(name="ContactInitialization")
 global_contact.initializationAssignments.appendInStep(
-    assignments=((
-    jaw_actuated_instance.surfaces['Surface'],
-    part_instance.surfaces['Surface'],
-    'ContactInitialization'), (
-    jaw_fixed_instance.surfaces['Surface'],
-    part_instance.surfaces['Surface'],
-    'ContactInitialization')), stepName='Initial')
+    assignments=(
+        (
+            jaw_actuated_instance.surfaces["Surface"],
+            part_instance.surfaces["Surface"],
+            "ContactInitialization",
+        ),
+        (
+            jaw_fixed_instance.surfaces["Surface"],
+            part_instance.surfaces["Surface"],
+            "ContactInitialization",
+        ),
+    ),
+    stepName="Initial",
+)
 
-#This makes the FEM run faster I think if it works
-#model.contactDetection(interactionProperty="IntProp-Fric", separationTolerance=0.0, extendByAngle=None,
+# This makes the FEM run faster I think if it works
+# model.contactDetection(interactionProperty="IntProp-Fric", separationTolerance=0.0, extendByAngle=None,
 #                       includeMeshShell=ON, includeMeshSolid=ON, includeMeshMembrane=ON,
 #                       includeNonOverlapping=OFF, createUnionOfMainSecondarySurfaces=ON, createUnionOfMainSurfaces=ON,
 #                       createUnionOfSecondarySurfaces=ON,
@@ -290,7 +313,8 @@ global_contact.initializationAssignments.appendInStep(
 #                       meshedGeometrySearchTechnique=USE_MESH)
 
 clamping_set = model.rootAssembly.Set(
-    name="Clamping-Load", referencePoints=(jaw_actuated_instance.referencePoints[2],))
+    name="Clamping-Load", referencePoints=(jaw_actuated_instance.referencePoints[2],)
+)
 
 model.ConcentratedForce(
     cf1=clamping_force,
@@ -321,31 +345,33 @@ mdb.jobs[job_name].writeInput()
 mdb.saveAs(str(job_name) + ".cae")
 job.submit()
 job.waitForCompletion()
-odb = visualization.openOdb(str(job_name) + ".odb")
-try:
-    frame = odb.steps["Apply-Force"].frames[-1]
-# noinspection
-except Exception, e:
-    file = open("retry.flag", "w")
-    file.close()
-    file = open("results.coam", "w")
-    file.write("0")
-    file.close()
-    sys.exit()
-
-dispField = frame.fieldOutputs["U"]
-max = 0
-for i in range(len(dispField.values)):
-    u1 = dispField.values[i].data[0]
-    max = u1 if u1 > max else max
-if max == 0:
+file = open("ClampSimulation.log", "r")
+if not "Abaqus JOB ClampSimulation COMPLETED" in file.read():
     # Try to fix with virtual topology if failed
     file = open("retry.flag", "w")
     file.close()
+    file = open("results.coam", "w")
+    file.write(json.dumps({"u1": 0, "s": 0}))
+    file.close()
+    sys.exit()
+
+odb = visualization.openOdb(str(job_name) + ".odb")
+frame = odb.steps["Apply-Force"].frames[-1]
+max_u1 = 0
+max_stress = 0
+
+dispField = frame.fieldOutputs["U"]
+for i in range(len(dispField.values)):
+    u1 = dispField.values[i].data[0]
+    max_u1 = u1 if u1 > max_u1 else max_u1
+
+dispField = frame.fieldOutputs["S"]
+for i in range(len(dispField.values)):
+    stress = dispField.values[i].data[0]
+    max_stress = stress if stress > max_stress else max_stress
 
 file = open("results.coam", "w")
-file.write(str(max))
+file.write(json.dumps({"u1": float(max_u1), "s": float(max_stress)}))
 file.close()
-
 
 # print(dispField)
